Commented, errors
Commented and made minor changes to some of my work. Read over all the code, and I honestly didnt know what was suposed to be error handled. Things like potentiometers cant really be error handled outside of applying a smoothing algortithem or smt. The only thing I could think of was the x/y bounds but I cant reallly test a fix from my grandmas place lol. I'll continue tommorow and maybe try to call in to work on a fix if need be.